lockdown - breakout making globalThis properties non-writable into lockdown-more.js
