+++ app/home.html
@@ -14,6 +14,7 @@
     <script src="./sentry-install.js" type="text/javascript" charset="utf-8"></script>
     <script src="./lockdown-install.js" type="text/javascript" charset="utf-8"></script>
     <script src="./lockdown-run.js" type="text/javascript" charset="utf-8"></script>
+    <script src="./lockdown-more.js" type="text/javascript" charset="utf-8"></script>
     <script src="./runtime-cjs.js" type="text/javascript" charset="utf-8"></script>
     {{@each(it.jsBundles) => val}}
     <script src="{{val}}" type="text/javascript" charset="utf-8"></script>
