+++ app/phishing.html
@@ -5,6 +5,7 @@
     <script src="./globalthis.js" type="text/javascript" charset="utf-8"></script>
     <script src="./lockdown-install.js" type="text/javascript" charset="utf-8"></script>
     <script src="./lockdown-run.js" type="text/javascript" charset="utf-8"></script>
+    <script src="./lockdown-more.js" type="text/javascript" charset="utf-8"></script>
     <script src="./phishing-detect.js"></script>
     <link rel="stylesheet" type="text/css" href="./index.css" title="ltr">
     <link rel="stylesheet" type="text/css" href="./index-rtl.css" title="rtl" disabled>
